CBA-17 Change type of Id columns

+++ src/main/java/com/thaichicken/cinemabooking/repository/CinemaHallRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CinemaHallRepository extends JpaRepository<CinemaHallEntity, Long> {
+public interface CinemaHallRepository extends JpaRepository<CinemaHallEntity, Integer> {
 }
