+++ src/main/java/com/thaichicken/cinemabooking/repository/ShowTimeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShowTimeRepository extends JpaRepository<ReservationEntity, Long> {
+public interface ShowTimeRepository extends JpaRepository<ReservationEntity, Integer> {
 }
